fixed all issues with basic c620 drive

+++ src/teensy/robot/c620.h
@@ -12,10 +12,10 @@
 
 class c620 : public rmMotor {
     private:
-        short id;
+        int id;
         CAN_message_t& sendMsg;
         int byteNum;
     public:
-        c620(short tempID, CAN_message_t* msg);
+        c620(int tempID, CAN_message_t* msg);
         void setPower(float power);
 };
